+++ sketch_oxide/src/similarity/mod.rs
@@ -18,6 +18,7 @@
 
 pub mod bbit_minhash;
 pub mod c_minhash;
+pub mod c_oph;
 pub mod lsh_ensemble;
 pub mod minhash;
 pub mod minhash_lsh;
@@ -31,6 +32,7 @@ pub mod weighted_minhash;
 
 pub use bbit_minhash::BBitMinHash;
 pub use c_minhash::CMinHash;
+pub use c_oph::COph;
 pub use lsh_ensemble::LshEnsemble;
 pub use minhash::MinHash;
 pub use minhash_lsh::MinHashLsh;